Normalising scatter plots with vmin, vmax
Changed the code to pass the vmin and vmax kwargs to the call to
ax.scatter() within scatter() in ternary/plotting.py i.e. vmin=vmin,
vmax=vmax. This allows the scatter plots to be normalised to the values
specified in the scatter() call. Without this, the min and max values of
the dataset are taken as defaults.