Changed offset behaviour of bottom_axis_label
Added a minus sign in the definition of position in bottom_axis_label()
so that increasing the value of offset moves the label down, further
below the bottom axis. This seemed more intuitive to me because when
increasingn the value of offset for the other two axes, the labels move
further away from the axes i.e. away from the centre of th simplex.