Moved function get_truncation
The function get_truncation was previously in one of the minor modules
but I have now moved it to ternary_axes_subplot as it makes much more
sence. This function has been rewritten and is now called
set_truncation, with get_truncation now simply returning
self._truncation, in alignment with other functions.